fix(html): also create pictures for plain.html
