+++ test/unit/rendering.test.js
@@ -150,6 +150,19 @@ describe('Rendering', () => {
       assert.ok(imgs.shift().getAttribute('loading') === 'eager', 'first image has loading set to eager');
       assert.ok(imgs.every((img) => img.getAttribute('loading') === 'lazy'), 'all other images have loading set to lazy');
     });
+
+    it('renders images.md correctly (plain)', async () => {
+      const html = await render('/images.md', 'plain_html');
+      const dom = new JSDOM(html);
+      const pics = Array.from(dom.window.document.querySelectorAll('picture'));
+      const imgs = pics.map((pic) => pic.querySelector('img'));
+      assert.strictEqual(pics.length, 3, 'document has 3 pictures');
+      assert.strictEqual(imgs.length, 3, 'document has 3 images');
+      assert.ok(pics.every((pic) => pic.querySelector('source').getAttribute('srcset').endsWith('?width=750&format=webply&optimize=medium')), 'pictures have source sets with correct parameters');
+      assert.ok(imgs.every((img) => img.getAttribute('src').endsWith('?width=2000&format=webply&optimize=medium')), 'images have source with correct parameters');
+      assert.ok(imgs.shift().getAttribute('loading') === 'eager', 'first image has loading set to eager');
+      assert.ok(imgs.every((img) => img.getAttribute('loading') === 'lazy'), 'all other images have loading set to lazy');
+    });
   });
 
   describe('Page Block', () => {
